Move DbContext helpers to core with exception processor
