+++ tilework.core/DbContextOptionsHelper.cs
@@ -1,9 +1,14 @@
 using Microsoft.EntityFrameworkCore;
+using EntityFramework.Exceptions.Sqlite;
+
+namespace Tilework.Core;
 
 public static class DbContextOptionsHelper
 {
     public static Action<DbContextOptionsBuilder> Configure(string connectionString)
     {
-        return options => options.UseSqlite(connectionString);
+        return options => options
+            .UseSqlite(connectionString)
+            .UseExceptionProcessor();
     }
 }
