+++ tilework.ui/Program.cs
@@ -5,6 +5,7 @@ using MudBlazor.Services;
 using Tilework.Ui.Components;
 using Tilework.Ui.ViewModels;
 
+using Tilework.Core;
 using Tilework.Core.Services;
 using Tilework.LoadBalancing.Services;
 using Tilework.CertificateManagement.Services;
